Fix source files arrangement
Fix folder name casing
Move logging code away from /scripts
Refactor existing code to fit new filepaths
