+++ src/analyzer/package_metadata_fetcher.py
@@ -1,19 +1,11 @@
-"""
-Creates PackageMetadata from a requirements.txt file
-"""
 import re
-import logging
 from infrastructure.pypi_client import PyPiHandler
 from entities.pypi_metadata import PyPiMetadata
 
-logger = logging.getLogger(__name__)
+_packagesmetadata: list[PyPiMetadata] = []
 
 
-
-# packagesmetadata: list[PyPiMetadata] = []
-
-
-def py_metadata_builder(file_path):
+def PyMetadataBuilder(file_path):
     """Reads a requirements.txt file and returns a list of dependencies.
 
         Args:
@@ -21,13 +13,7 @@ def py_metadata_builder(file_path):
         Returns:
             list (str): A list of dependencies specified in the file.
         """
-    # Tiro giù tutto l'albero con pipdeptree e poi verifico
-    # per ognuna repo vs pipy.
-    # Dopodichè avrò raccolto tutte le licenze e proseguirò a fare un
-    # confronto di compatibilità
-    # SOLO DEL REPO IN ESAME VS QUELLO
-    # noqa: E501 If we will pass the whole tree this should be global
-    packagesmetadata: list[PyPiMetadata] = []
+    # packagesmetadata: list[PyPiMetadata] = []  # noqa: E501 If we will pass the whole tree this should be global
 
     dependencies = []
     pattern = re.compile(r"^\s*([A-Za-z0-9_.-]+)")
@@ -40,34 +26,40 @@ def py_metadata_builder(file_path):
                     dependencies.append(match.group(1))
 
     except FileNotFoundError:
-        logger.error("File not found: %s ", file_path)
-    except IOError as e:
-        logger.error("An I/O error occurred: %s", e)
+        print(f"File not found: {file_path}")
+        return []
+    except OSError as e:
+        print(f"Error reading file {file_path}: {e}")
+        return []
+    except Exception as e:  # pylint: disable=broad-exception-caught
+        print(f"Unexpected error parsing {file_path}: {e}")
+        return []
 
+
+    # Pass requirement to treebuilder to get full dependency tree
     results = PyPiHandler.get_source_links(dependencies)
 
     for pkg_name, metadata in results.items():
-        packagesmetadata.append(PyPiMetadata(
+        _packagesmetadata.append(PyPiMetadata(
             package=pkg_name,
-            license_name=metadata['license'],
+            license_type=metadata['license'],
             link=metadata['link']
         ))
+        
+    # The process should stop here. The user then could click on a dependency on the interactive tree to see its details.
+    # At this point the user could launch a check license from source code (ScanCode part)
+    # Finally an option "check all tree licenses from source code" outside of the tree 
+    # will launch the Downlaod and ScanCode partfor all dependencies.
 
-    # calls PyPiClient and retrieves links, constructing a
-    # PyPiMetadata object for each package
-    # [question -> depth-first or breadth-first search?
-    # Depth-first in my opinion UPDATE!! I've found pipdeptree]
+    # calls PyPiClient and retrieves links, constructing a PyPiMetadata object for each package
     # then calls RepoDownloader and downloads the repo for each package
     # then calls ScanCodeRunner and creates ScanCodeResults for each repo
-    # given a package and its (PyPiMetadata, ScanCodeResults) pair, create
-    # PackageMetadata and check that the two licenses match
-    # at this point, analyze whether all dependency licenses are compatible
-    # with the license of the project being analyzed
+    # given a package and its (PyPiMetadata, ScanCodeResults) pair, create PackageMetadata and check that the two licenses match
+    # at this point, analyze whether all dependency licenses are compatible with the license of the project being analyzed
 
-    # ↓ I'm not pratical with python but we should ensure
-    # encuplation in some way
-    return packagesmetadata
+    # ↓ I'm not pratical with python but we should ensure encuplation in some way
+    return _packagesmetadata
 
 
-def pypi_license_checker():
-    """Checks the license of a PyPI-hosted package"""
+def PyPiLicenseChecker():
+    pass
