+++ src/entities/package_metadata.py
@@ -3,4 +3,4 @@
 
 class PackageMetadata:
     """Contains metadata about a package"""
-    pass
+
